Stairs fixed.  More loot in pawn shops.

+++ line.cpp
@@ -60,7 +60,7 @@ std::vector <point> line_to(int x1, int y1, int x2, int y2, int t) {
 
 int trig_dist(int x1, int y1, int x2, int y2)
 {
- return int(sqrt(pow(x1 - x2, 2) + pow(y1 - y2, 2)));
+ return int(sqrt(double(pow(x1 - x2, 2.0) + pow(y1 - y2, 2.0))));
 }
 
 int rl_dist(int x1, int y1, int x2, int y2)
